+++ lib/amp_sdk/transport.ex
@@ -6,12 +6,11 @@ defmodule AmpSdk.Transport do
   @type t :: pid()
   @type message :: binary()
   @type opts :: keyword()
-  @type subscription_tag :: :legacy | reference()
+  @type subscription_tag :: reference()
 
   @callback start(opts()) :: {:ok, t()} | {:error, term()}
   @callback start_link(opts()) :: {:ok, t()} | {:error, term()}
   @callback send(t(), message()) :: :ok | {:error, term()}
-  @callback subscribe(t(), pid()) :: :ok | {:error, term()}
   @callback subscribe(t(), pid(), subscription_tag()) :: :ok | {:error, term()}
   @callback close(t()) :: :ok
   @callback force_close(t()) :: :ok | {:error, term()}